Cleanup previous awards for a given route on a given date on a given engine before saving updated awards data.
