+++ bin/cli-search.js
@@ -164,8 +164,8 @@ function generateQueries (args, engine, days) {
   // Get search strategy based on engine
   const {
     roundtripOptimized = true,
-    oneWaySupported = true,
-    tripMinDays = 3
+      oneWaySupported = true,
+      tripMinDays = 3
   } = strategies[engine.id.toLowerCase()] || {}
   const gap = (args.oneway || !roundtripOptimized) ? 0 : Math.min(tripMinDays, days)
   const validEnd = engine.config.validDateRange()[1]
@@ -353,7 +353,7 @@ const main = async (args) => {
 
       // Lazy load the search engine
       if (!initialized) {
-        const credentials = loginRequired
+        const credentials = loginRequired 
           ? accounts.getCredentials(id, args.account) : null
         await engine.initialize({ ...options, credentials })
         initialized = true
@@ -399,7 +399,7 @@ const main = async (args) => {
           daysRemaining = terminate // Reset termination counter
         }
         const placeholders = helpers.createPlaceholders(results, { cabins: Object.values(fp.cabins) })
-        helpers.saveAwards(requestId, awards, placeholders)
+        helpers.saveAwards(requestId, awards, placeholders, query)
       }
     }
     if (skipped > 0) {
